fix: use for all term commands
