+++ helix-view/src/lib.rs
@@ -4,6 +4,7 @@ pub mod macros;
 pub mod clipboard;
 pub mod document;
 pub mod editor;
+pub mod env;
 pub mod graphics;
 pub mod gutter;
 pub mod handlers {
